+++ sketch-compose-resources/src/jvmCommonTest/kotlin/com/github/panpf/sketch/compose/resources/jvm/test/fetch/internal/ComposeResourceUriFetcherComponentJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.compose.resources.jvm.test.fetch.internal
+
+class ComposeResourceUriFetcherComponentJvmTest {
+    // TODO test
+}