+++ sketch-compose-resources/src/nonJvmCommonTest/kotlin/com/github/panpf/sketch/compose/resources/nonjvm/test/fetch/internal/ComposeResourceUriFetcherComponentNonJvmTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.compose.resources.nonjvm.test.fetch.internal
+
+class ComposeResourceUriFetcherComponentNonJvmTest {
+    // TODO test
+}